added all reworked classes

+++ shared/SimulationSettingsData.cs
@@ -2,7 +2,7 @@ namespace shared;
 
 public class SimulationSettingsData
 {
-    public int SimulationSpeed { get; set; }
+    public float SimulationSpeed { get; set; }
     public float TrainWaitingTimeAtStation { get; set; }
     public List<string> LoadingStationIds { get; set; } = [];
     public List<string> ChargingStationIds { get; set; } = [];
